This time, fluid recipes and creosote bottles are all right!
Also start to meddle with JEI

+++ src/main/java/mods/railcraft/common/fluids/AdvancedFluidHandler.java
@@ -82,7 +82,7 @@ public class AdvancedFluidHandler implements IFluidHandler {
             amount += liquid == null ? 0 : liquid.amount;
             capacity += tank.getCapacity();
         }
-        return capacity == 0 ? 0 : amount / capacity;
+        return capacity == 0 ? 0 : ((float) amount) / capacity;
     }
 
     public float getFluidLevel(Fluid fluid) {
